refactor(i18n): consolidate strings.ts into i18n.ts
- Move all locale loading logic from strings.ts into i18n.ts
- Delete strings.ts — i18n.ts is now the canonical locale module
- Update prompts.ts and prompts-data.ts to import Lang from ./i18n
